+++ src/services/ImageToolboxService.php
@@ -22,6 +22,8 @@ use Twig\Error\RuntimeError;
 use craft\elements\Asset;
 
 
+
+
 /**
  * @author    Piotr Pogorzelski
  * @package   ImageToolbox
@@ -220,6 +222,90 @@ class ImageToolboxService extends Component
         return false;
     }
 
+    // copied from ImageTransforms class
+
+    public static function generateTransformImageObj(
+        Asset $asset,
+        \craft\models\ImageTransform $transform,
+        // ?callable $heartbeat = null,
+        // ?BaseImage &$image = null,
+    // ): string {
+    ) {
+        $ext = strtolower($asset->getExtension());
+        // if (!Image::canManipulateAsImage($ext)) {
+        //     throw new ImageTransformException("Transforming .$ext files is not supported.");
+        // }
+
+        // $format = $transform->format ?: static::detectTransformFormat($asset);
+        $imagesService = Craft::$app->getImages();
+
+        // $supported = match ($format) {
+        //     Format::ID_WEBP => $imagesService->getSupportsWebP(),
+        //     Format::ID_AVIF => $imagesService->getSupportsAvif(),
+        //     Format::ID_HEIC => $imagesService->getSupportsHeic(),
+        //     default => true,
+        // };
+
+        // if (!$supported) {
+        //     throw new ImageTransformException("The `$format` format is not supported on this server.");
+        // }
+
+        // $generalConfig = Craft::$app->getConfig()->getGeneral();
+        $imageSource = \craft\helpers\ImageTransforms::getLocalImageSource($asset);
+
+        if ($ext === 'svg' && $format !== 'svg') {
+            $size = max($transform->width, $transform->height) ?? 1000;
+            $image = $imagesService->loadImage($imageSource, true, $size);
+        } else {
+            $image = $imagesService->loadImage($imageSource);
+        }
+
+        // if ($image instanceof Raster) {
+        //     $image->setQuality($transform->quality ?: $generalConfig->defaultImageQuality);
+        //     $image->setHeartbeatCallback($heartbeat);
+        // }
+
+        if ($asset->getHasFocalPoint() && $transform->mode === 'crop') {
+            $position = $asset->getFocalPoint();
+        } elseif (!preg_match('/^(top|center|bottom)-(left|center|right)$/', $transform->position)) {
+            $position = 'center-center';
+        } else {
+            $position = $transform->position;
+        }
+
+        $scaleIfSmaller = $transform->upscale ?? Craft::$app->getConfig()->getGeneral()->upscaleImages;
+
+        switch ($transform->mode) {
+            case 'letterbox':
+                if ($image instanceof \craft\image\Raster) {
+                    $image->scaleToFitAndFill(
+                        $transform->width,
+                        $transform->height,
+                        $transform->fill,
+                        $position,
+                        $scaleIfSmaller
+                    );
+                } else {
+                    // Craft::warning("Cannot add fill to non-raster images");
+                    $image->scaleToFit($transform->width, $transform->height, $scaleIfSmaller);
+                }
+                break;
+            case 'fit':
+                $image->scaleToFit($transform->width, $transform->height, $scaleIfSmaller);
+                break;
+            case 'stretch':
+                $image->resize($transform->width, $transform->height);
+                break;
+            default:
+                $image->scaleAndCrop($transform->width, $transform->height, $scaleIfSmaller, $position);
+        }
+
+        // if ($image instanceof Raster) {
+        //     $image->setInterlace($transform->interlace);
+        // }
+
+        return $image;
+    }
 
     private static function getWidthHeightAttrs($asset, $transform)
     {   
@@ -243,29 +329,17 @@ class ImageToolboxService extends Component
         if(!isset($transform['width']) && !isset($transform['height']) && !is_null($asset)){
             return [
                 'width' => $asset->width,
-                'height' => $asset->width,
+                'height' => $asset->height,
             ];            
         }
 
-        // if both width height transform
-        if(isset($transform['width']) && isset($transform['height'])){
-            $upscale = Craft::$app->getConfig()->getGeneral()->upscaleImages;
-            // if upscaling enabled or no asset and we can make placeholder any size we want
-            if($upscale == true || is_null($asset)){
-                return [
-                    'width' => $transform['width'],
-                    'height' => $transform['height'],
-                ];
-            }else{
-                return [
-                    'width' => $transform['width'] > $asset->width ? $asset->width : $transform['width'],
-                    'height' => $transform['height'] > $asset->height ? $asset->height : $transform['height'],
-                ];                
-            }
-        }
-
-        // if image exists but only height or only width in transform
-        return null;
+        // regular transform
+        $transformSettings = new \craft\models\ImageTransform($transform);
+        $image = self::generateTransformImageObj($asset, $transformSettings);
+        return [
+            'width' => $image->width,
+            'height' => $image->height,
+        ];
 
     }
 
@@ -345,6 +419,11 @@ class ImageToolboxService extends Component
         $fallback_attributes = [
             'src' => $fallback_src,
         ];
+
+        if(!is_null(self::getWidthHeightAttrs($image, $fallback_transform))){
+            $fallback_attributes['width'] = self::getWidthHeightAttrs($image, $fallback_transform)['width'];
+            $fallback_attributes['height'] = self::getWidthHeightAttrs($image, $fallback_transform)['height'];
+        }
 
         // add provided attributes
         if(!is_null($attributes)){